refactor(tui): split page_fleet.go god-methods into feature files
Reduce page_fleet.go to the page spine (fleetPage struct, constructor,
Init/Update/View dispatchers) and move the rest into focused files:
- page_fleet_rows.go: buildRows, cursor/nav/selection, focus mode
- page_fleet_keys.go: updateNormal, handleEnter, contextual help
- page_fleet_view.go: viewFleetList, viewActiveDialog dispatcher, armada border
- page_fleet_split.go: split-pane open + session-group restore/cycle
openArmadaSelect folds into armada.go; renderLayoutPresetOverlay into
layout_preset_dialog.go. Pure code motion; tidy stale field names in test
messages. No behavior change.
